Add refactor onDelete fn
